pyproject.toml
[tool.poetry]
name = "django-vercel"
version = "0.1.0"
description = ""
authors = ["Arvind-4 <lordarvindthegamer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.14"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
whitenoise = "^6.1.0"
django-dotenv = "^1.4.2"
django-ckeditor = "^6.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
